Fix async tweet-bot call
When aws runs a lambda kicked off by an event it is run asynchronous. This means nothing is waiting for the output, when the handler function finishes the lambda vm is destroyed.
As we did not await the `saveTweetedPost` call with its chained then parts the vm finishes successfully and gets destroeyed.

The then-catch logic was replaced with await to make it more readable and some logging was added to make it debugging with aws logs easier.