Do not expose axios
A client like this one for moco should abstract away the api. Code using this
function should not need to know how it is implemented or works internally.
This includes the use of axios. When i use this function i do not need to know
this axios is used for making a http request.
That is why the function should not return a axios response and just return the
template.