Remove unused fast-apply handler
Can not find any usage inside serverless.yml. Based on the git history and gitlab search (https://gitlab.com/newcubator/newcubator/micro-bots/-/commits/main?ref_type=heads?search=fast-apply) it looks like it was never used. When using the form on newcubator.com the result is send to https://kceiclszhf.execute-api.eu-central-1.amazonaws.com/production/contactRequest which is a different function. Based on all of this I asume the code is unsued and can be removed.