Fix and enable eslint
Eslint was not run in ci until now. This lead to a lot of errors on main. This MR updates the eslint config to the new format and enables eslint in ci.
Since fixing all current errros would be a to big MR all CURRENT errors are ignored as warnings for now. Idea is to fix them bit by bit in subsequent MRs.